apps/api/requirements.txt
# CreditBeast API Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
supabase>=2.9.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
# clerk-backend-api>=3.3.0  # Temporarily disabled due to dependency conflicts

# Stripe integration
stripe==7.10.0

# Email (CloudMail placeholder - will use SMTP)
aiosmtplib==3.0.1
jinja2==3.1.3

# Utilities
httpx>=0.26.0
python-dateutil==2.8.2
pytz==2024.1

# Export & Reporting
reportlab==4.0.7  # PDF generation
pandas==2.1.4  # CSV and data manipulation
cryptography==43.0.3  # For PII encryption (already imported in database.py)

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
